requirements.txt
pyyaml
TurboGears2
beaker
pylgtv
denonavr
wakeonlan